backend/requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
blinker==1.6.2
boto3==1.28.30
botocore==1.31.30
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
distlib==0.3.6
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.12.2
Flask==2.3.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
legacy==0.1.6
MarkupSafe==2.1.2
packaging==23.2
Pillow==10.1.0
platformdirs==3.5.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pybase62==1.0.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uuid0==0.2.7
virtualenv==20.23.0
Werkzeug==2.3.4
